From: Colm Dougan <colm.dougan@replify.com>
Problem with list_rules_IPs() on 64-bit platforms... fixed by Colm Dougan.


I'm using this module on Debian etch and am getting a segfault with
the following code :

my $tname = 'nat';
my $cname = 'PREROUTING';

my $table = IPTables::libiptc::init($tname);

foreach my $rule ($table->list_rules_IPs('src', $cname)) {
    print "Src => ", Dumper(\$rule), "\n";
}

Here is a patch :

diff -Brub IPTables-libiptc-0.14_orig/libiptc.xs
IPTables-libiptc-0.14/libiptc.xs
--- IPTables-libiptc-0.14_orig/libiptc.xs       2008-05-26
11:54:37.000000000 +0100
+++ IPTables-libiptc-0.14/libiptc.xs    2009-07-09 14:48:03.000000000 +0100
@@ -5,6 +5,7 @@
 #include "ppport.h"

 #include <libiptc/libiptc.h>
+#include <include/iptables.h>
 #include <string.h>
 #include <stdio.h>
 #include <errno.h>

The problem was that libiptc.xs had no prototype for 'addr_to_dotted'
and so it was defaulting to an 'int' which luckily did the right thing
on 32bit architectures but the wrong thing on 64 bit.

I was also thinking it might be worth compiling with -Werror and -Wall
so that this type of thing gets caught at compile time.  However there
would be a few fixups that would need to be made to the code to
compile clean with those options.  I can have a look at that if you
think it is worthwhile (although I can't promise to do it right away).