Check that the xtables.h header is available on the system
Add functionality to check if a C header file is available
on the system.  The bash script "check-header", will create
a very small C progream which include the given header file
and then try to compile the program using gcc.

This functionality is used for checking if the header file
"xtables.h" exists on the system.  If not, a very very
verbose message is written to the user, when he calls
perl Makefile.PL.

Signed-off-by: Jesper Dangaard Brouer <hawk@comx.dk>