Disable locking in iptables library
iptables before 1.6.0 added a fifth argument to do_command4() to
control locking. Perl module never passed the argument and thus
invoked in undefined behavior.

This become visible in iptables-1.8.0 that changed the locking and
that caused an infinite loop in Perl iptables_do_command(). The
iptables locking mechanmism is not firendly for libraries, it can call
exit(2).

This this patch fixes the infinite loop by disabling locking
explicitly.

Red Hat Bugzilla #1670047 related to Fedora 26,28,29 and 30
 https://bugzilla.redhat.com/show_bug.cgi?id=1670047

Patch found via Fedora package:
 https://src.fedoraproject.org/rpms/perl-IPTables-libiptc

Signed-off-by: Petr Písař <ppisar@redhat.com>
Signed-off-by: Jesper Dangaard Brouer <netoptimizer@brouer.com>